Only attempt to delete if the file exists

+++ spec/parser_spec.rb
@@ -14,7 +14,7 @@ describe Sawyer::Parser, "#now" do
   end
 
   it "parses lines correctly" do
-    File.delete(@offset_file)
+    File.delete(@offset_file) if File.exist?(@offset_file)
     parser = Sawyer::Parser.new(@logfile, @offset_file)
     info_metric = 'sawyer.info'
     warn_metric = 'sawyer.warn'
